Add i9-7900x benchmark?
I have no idea why but it seems like an i9-7900x is just... really bad at
PSADBW/VPSADBW for some reason. Gonna investigate this later and add it to the
readme just in case there is some weird latency thing I'm not aware of for
Skylake-X's vector units or something.