x64: Implement full 16x16 AMX multiplication
Multiply a 4x16 vector of masks against a full 16x16 tile of pixel-data to increase through-put.
The local-to-global sum routine at the end of the loop is still due for a rework
and I'm not _too_ sure about the `LocalSumOverflowMax` calculation in this case
but this passes all of the unit tests.
The AMX implementation now handles 256(16 * 16)-pixels per iteration.